Add test to detect statement in evaluation.
This commit detect when a statement is evaluated with PyRun_String run
with Py_eval_input. This seems to generate a Syntax Error. Thus if
an exception with Syntax Error is generated with Py_eval_input,
the python interpreter does a second pass with Py_single_input.
The string err is set appropriatly, out has to be checked.